refactor: extracting component
