+++ src/components/moleculars/StatisticsCard/Icon.tsx
@@ -1,0 +1,15 @@
+import TicketOutlinedIcon from "components/vectors/TicketOutlinedIcon";
+import RibonFlagIcon from "components/vectors/RibonFlagIcon";
+import { IconType } from ".";
+
+type IconProps = { icon: IconType };
+export function Icon({ icon }: IconProps) {
+  switch (icon) {
+    case "TicketIconOutlined":
+      return <TicketOutlinedIcon />;
+    case "RibonFlagIcon":
+      return <RibonFlagIcon />;
+    default:
+      return null;
+  }
+}
